Update Yarn to 1.0.1
Summary:
  - Allows the top-level package.json to override nested dependencies versions
  - Adds a yarn-path rc config parameter
  - Enables Workspaces by default
  - Commands are now ran from the cwd package root if possible
  - Exposes the lockfile parser into its own npm package (@yarnpkg/lockfile)
  - Implements a --cwd flag to run Yarn in a specific directory
  - Detects and merges lockfile conflicts without user input
  - Adds a preliminary support for // comments in package.json dependencies

The patch that was being applied can be removed as the fix is present in this release. Also, the gulp build stage no longer has a 'build-legacy' step, so the directory 'lib-legacy' is no not needed to be copied across during the install stage.

Test Plan: Verified the cli was working by executing yarn inside a project that uses it.

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Tags: #patch_submission

Differential Revision: https://dev.solus-project.com/D976